Test new preference/mapPreferences implementation within a vue component
