+++ tests/integration/install.spec.js
